install: don't copy /etc/nginx/sso if it already exists

+++ install.py
@@ -259,8 +259,9 @@ def install_nginxcfg():
                 mode=0o644)
     install_cfg('nginx/enable_sso', '/etc/nginx', owner='root:root',
                 mode=0o644)
-    copytree('etc/nginx/sso', '/etc/nginx/sso', owner='http:root',
-             dir_mode=0o750, file_mode=0o640)
+    if not os.path.exists('/etc/nginx/sso'):
+        copytree('etc/nginx/sso', '/etc/nginx/sso', owner='http:root',
+                 dir_mode=0o750, file_mode=0o640)
     mkdir('/etc/nginx/services', mode=0o755, owner='root:root')
     mkdir('/etc/nginx/services_contest', mode=0o755, owner='root:root')
     if not os.path.exists('/etc/nginx/logs'):
